Add voice prompt experiment
You'll need pocketsphinx:
* brew install sphinxbase --HEAD
* brew install pocketsphinx --HEAD

The doing `alexa ask -d <pocketsphinx dir in this repo> -p` will cause
ask to wait for you to say 'Hey alexa' and then the sentence after that
will be sent to Alexa!

It even runs in a loop, so you could leave it running all the time!